Create a build_command.py file (#375)
